refactoring directory structure
As the app will get bigger and have more features, it's crucial to separate code correcty. I opted for a feature-first approach, with a single entry point being main.dart (this will be useful later when routing will be added, we will just import the necessary screens and have it render them).