add go_router
For the architecture I want to build, it is crucial to use a routing package to be able to import everything in main.dart and have the full app structure visible from the entry point. I think that would make it easier to undertand and also easier to debug. go_router also comes with a lot of great features, such as being able to change the transition animation, create a catch-all error page and much more.