fix redirect to work with AuthProvider
Turns out I need some sleep. AuthProvider was not redirecting because the redirect clause was inside  the route itself, not as global. Now, when auth state changes, it will either redirect to / or /login if AuthenticationProvider notifies listeners, which is the two possible routes we could go to (either user logs in / signs up and gets in or user logs off and  gets sent back to login screen)