refactor error handling
I refactored a lot of the error handling in AuthProvider, as the status was a bit awkward  to iterate over. I decided to take advantage of Dart's typing by checking directly the type of a generic sealed class that is overriden by more specific implementations. This allows us to have way nicer error handling, as well as a more generic interface.